Refatoração concluída, separado arquivos e etc.
