+++ src/models/Vertice.ts
@@ -1,0 +1,9 @@
+export class Vertice {
+  coordenada: string;
+  pai: Vertice | null;
+
+  constructor(coordenada: string, pai: Vertice | null) {
+    this.coordenada = coordenada;
+    this.pai = pai;
+  }
+}
